+++ packages/webpack-test/tests/compile-module-date.test.ts
@@ -61,7 +61,7 @@ const getWebpackDefaults = (PROJECT_DIR: string, OUTPUT_DIR: string) => ({
 });
 
 beforeAll(() => {
-    fs.rmSync(path.resolve(path.join(__dirname, "..", "..", "example-addons", "lib")), { recursive: true, force: true });
+    fs.rmSync(path.resolve(__dirname, "..", ".websmith-cache", "addons"), { recursive: true, force: true });
 });
 
 let PROJECT_DIR: string;